feat: seperate jvectormap & fix some issue
