+++ src/assets/js/maps/index.js
@@ -1,1 +1,2 @@
 export * from './google-map';
+export * from './vector-map';
